clean up start server result types
